KFS::config::load_config(...) now checks first if filepath is file or dir, and then tries to load

+++ config.py
@@ -22,32 +22,39 @@ def load_config(filepath: str, default_content: str="", empty_ok: bool=False) ->
     else:                                       #if no root logger defined:
         logger=log.setup_logging("KFS")         #use KFS default format
         
-
-    logger.info(f"Loading \"{filepath}\"...")
-    try:
-        with open(filepath, "rt") as file:  #read file
-            file_content=file.read()
-    except FileNotFoundError:   #if file does not exist:
+    if os.path.isfile(filepath)==False and os.path.isdir(filepath)==False:  #if input configuration file does not exist yet: create a default one and print instructions
+        logger.warning(f"Did not load \"{filepath}\", because file does not exist.")
         file_content=None
-        logger.warning(f"Could not load \"{filepath}\", because file does not exist.")
-        logger.info(f"Creating \"{filepath}\" and filling with default content...")
+        logger.info(f"Creating default \"{filepath}\"...")
         try:
             if os.path.dirname(filepath)!="":
                 os.makedirs(os.path.dirname(filepath), exist_ok=True)   #create folders
             with open(filepath, "wt") as file:                          #create file
                 file.write(default_content)                             #fill with default content
         except OSError:
-            logger.error(f"Could not create \"{filepath}\" and fill with default content.")
+            logger.error(f"Creating default \"{filepath}\" failed.")
         else:
-            logger.info(f"\rCreated \"{filepath}\" and filled with default content.")
-    except PermissionError: #if filepath is actually a directory:
+            logger.info(f"\rCreated default \"{filepath}\".")
+        return file_content
+    elif os.path.isfile(filepath)==False and os.path.isdir(filepath)==True:
+        logger.error(f"Did not load \"{filepath}\", because it is a directory. Unable to create default file.")
         file_content=None
-        logger.error(f"Could not load \"{filepath}\", because permission was denied. It is likely a folder and not a file.")
-    else:                                                                   #if loading successful:
-        if file_content=="" and empty_ok==False:                            #but if content is empty and not allowed empty:
-            logger.warning(f"\rLoaded \"{filepath}\", but it is empty.")    #warning
-            file_content=None
-        else:
-            logger.info(f"\rLoaded \"{filepath}\".")
+        return file_content
+
+
+    logger.info(f"Loading \"{filepath}\"...")
+    try:
+        with open(filepath, "rt") as file:  #read file
+            file_content=file.read()
+    except OSError: #if filepath is actually a directory:
+        logger.error(f"Loading \"{filepath}\" failed.")
+        file_content=None
+        return file_content
+    
+    if file_content=="" and empty_ok==False:                        #but if content is empty and not allowed empty:
+        logger.error(f"\rLoaded \"{filepath}\", but it is empty.")  #error
+        file_content=None
+    else:
+        logger.info(f"\rLoaded \"{filepath}\".")
     
     return file_content
